Delta Time 적용 (시간 동기화)

+++ 2022_05_12_ShootingGame/Bullet.h
@@ -37,8 +37,8 @@ protected:
 	Vector2		destroyImgRectSize;
 	POINT		destroyFrame;
 
-	int			frameTime;
-	int			destroyTime;
+	double		frameTime;
+	double		destroyTime;
 
 private:
 	// Debug Object
